github-patches-check: xmastree: make @@ handling match upstream
The upstream xmastree.py[1] code does not (and as far as I can tell)
has never included a continue at the end of the @@ handling block.

Removing it from our local version, which was inherited from
nfp-drv-kmods[2] resolves a false positive where by the parameters
of a function declaration were flaged for an xmastree violation (the rule
does not apply in that context).

Hunk:

@@ -1917,12 +1917,14 @@ static int fl_set_parms(struct net *net, struct tcf_proto *tp,
                        struct cls_fl_filter *f, struct fl_flow_mask *mask,
                        unsigned long base, struct nlattr **tb,
                        struct nlattr *est,
-                       struct fl_flow_tmplt *tmplt, u32 flags,
+                       struct fl_flow_tmplt *tmplt,
+                       u32 flags, u32 fl_flags,
                        struct netlink_ext_ack *extack)
 {
        int err;

Warning:

Line 1920
	struct nlattr *est,
	struct fl_flow_tmplt *tmplt,
Line 1922
	u32 flags, u32 fl_flags,
	struct netlink_ext_ack *extack)

[1] https://github.com/ecree-solarflare/xmastree/
[2] https://github.com/Corigine/nfp-drv-kmods-private/

Signed-off-by: Simon Horman <simon.horman@corigine.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@corigine.com>