github-patches-check: Add check for newline at end of message in NL_SET_ERR_MSG_MOD
Currently coccicheck warnings are ignored and patches have been sent
that adds trivial mistakes upstream that had to be fixed by third
parties, see for example [1].

In an ideal world all warnings reported by coccicheck should fail the
commit check. Before that can be enforced all current warnings needs to
be fixed. To avoid adding more warnings of the type in [1] add a specific
check for it.

1. commit 0c1794c200e943ea ("nfp: xsk: avoid newline at the end of message in NL_SET_ERR_MSG_MOD"

Signed-off-by: Niklas Söderlund <niklas.soderlund@corigine.com>