github-patches-check: fetch full git tree if needed by checkpatch
Some checks in checkpatch.pl require access to the full git worktree.
The default behavior of the workflow is to only do a shallow clone to
save on resources. This can lead to hard to debug checkpatch errors as
running the tests locally passes but fails in the workflow.

To avoid this add a test if checkpatch will need access to the full
worktree before starting to check individual commits, and if so
unshallow the worktree.

Signed-off-by: Niklas Söderlund <niklas.soderlund@corigine.com>