github-patches-check: allow all checks to run before failing
Creating the environment to test commits in are expensive but the
current commit check script terminates the job at the first error. This
can be inefficient as multiple small errors may exists but only be
revealed one at a time.

Change the structure of how a commit it tested to allow multiple
different checks to run before potentially failing the commit and halt
the verification job. This allow all the tests to try and find as many
faults as possible while still failing the commit if just a single check
step fails.

Signed-off-by: Niklas Söderlund <niklas.soderlund@corigine.com>