github-patches-check: fix and not always do an unshallow clone
Avoid "fetch full git tree" always triggers, even when not needed,
as commit 4af6bacfab (github-patches-check: Add check for newline
at end of message in NL_SET_ERR_MSG_MOD) contains a reference to a
commit.

This adds quite an significant amount of unnecessary time in a lot
of cases where there are no Commit of Fixes tag in any of the actual
patches.

Signed-off-by: Yu Xiao <yu.xiao@corigine.com>