github-patches-check: ignore test workflow patches for unshallow decision
Since the check to unshallow the tree when needed a patch that triggers
the decision have been added to the github workflow patches themself.
This leads to that the tree us always unshallowed.

Avoid this by looking at each commit individually and ignore the github
workflow patches. This still needs to happen as a separate pre step to
avoid touching all the files in the worktree once the checking of the
real commits have begun.

Signed-off-by: Niklas Söderlund <niklas.soderlund@corigine.com>