github-patches-check: fix up unshallow clone update
The previous patch had some bugs that now caused the opposite
problem - it always skips the 'git fetch --unshallow' command,
even when it shouldn't. Luckily it only requires a few small
changes to get this fixed - the git commands need to reference
the specific commit variable which is being iterated.

Signed-off-by: Louis Peens <louis.peens@corigine.com>